Move absolute URL utils to hourglass.site_utils.
