+++ data_capture/email.py
@@ -1,14 +1,13 @@
 import re
 import premailer
-from django.contrib.sites.models import Site
 from django.core.mail import EmailMultiAlternatives, get_connection
-from django.core.urlresolvers import reverse
 from django.utils.html import strip_tags
 from django.template.loader import render_to_string
 from django.contrib.auth.models import User
 from django.template.defaultfilters import pluralize
 from django.conf import settings
 
+from hourglass.site_utils import absolute_reverse
 from .models import SubmittedPriceList
 
 
@@ -67,36 +66,6 @@ def render_mail(template, ctx):
     )
 
     return (plaintext_message, html_message)
-
-
-def absolutify_url(url):
-    '''
-    If the URL is an absolute path, returns the URL prefixed with
-    the site's protocol and host information.
-
-    If the given URL is already absolute, returns the URL unchanged.
-    '''
-
-    if url.startswith('http://') or url.startswith('https://'):
-        return url
-
-    if not url.startswith('/'):
-        raise ValueError(f"url must be an absolute path: {url}")
-
-    protocol = 'https'
-    host = Site.objects.get_current().domain
-    if settings.DEBUG or not settings.SECURE_SSL_REDIRECT:
-        protocol = 'http'
-    return f"{protocol}://{host}{url}"
-
-
-def absolute_reverse(*args, **kwargs):
-    '''
-    Like Django's reverse(), but ensures the URL includes protocol
-    and host information, so that it can be embedded in an email.
-    '''
-
-    return absolutify_url(reverse(*args, **kwargs))
 
 
 def send_mail(subject, to, template, ctx, reply_to=None):
